==> 04_ScenarioPartitioning.Rmd <==
---
title: "AnnAGNPS Data Organization Pt IV"
author: "Brady Nahkala"
date: "9/1/2020"
output: html_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
```

This is part IV of my data processing organization where I start to look at the data by scenario or common scenario groupings. 

# Library 

```{r}
library(dplyr)
library(tidyr)
library(reshape2)
library(ggplot2)
library(ggpubr)
```

# Naming and numbering conventions

```{r}
simulation_numbers <- data.frame("Drain0" = c(22, 23, 24, 3, 12, 13, 1),
                                 "Drain1" = c(4, 14, 25, 19, 29, 33, 36),
                                 "Drain2" = c(5, 15, 26, 20, 30, 34, 37),
                                 "Drain3" = c(6, 16, 27, 21, 31, 35, 38))
simulation_numbers2 <- data.frame("CSCV" = c(22, 4, 5, 6),
                                 "CSCT" = c(23, 14, 15, 16),
                                 "CSNT" = c(24, 25, 26, 27),
                                 "RCV" = c(3, 19, 20, 21),
                                 "RCT" = c(12, 29, 30, 31),
                                 "RNT" = c(13, 33, 34, 35),
                                 "RR" = c(1, 36, 37, 38))

simulation_numbers3 <- data.frame("PDCC" = c(0, 7, 8, 9, 10, 17))

simulation_shortIDs <- data.frame("Drain0" = c("D0-CS-CV", "D0-CS-CT", "D0-CS-NT", "D0-R-CV", "D0-R-CT", "D0-R-NT","D0-R-R"),
                                  "Drain1" = c("D1-CS-CV", "D1-CS-CT", "D1-CS-NT", "D1-R-CV", "D1-R-CT", "D1-R-NT","D1-R-R"),
                                  "Drain2" = c("D2-CS-CV", "D2-CS-CT", "D2-CS-NT", "D2-R-CV", "D2-R-CT", "D2-R-NT","D2-R-R"),
                                  "Drain3" = c("D3-CS-CV", "D3-CS-CT", "D3-CS-NT", "D3-R-CV", "D3-R-CT", "D3-R-NT","D3-R-R"))
simulation_shortIDs2 <- data.frame("CSCV" = c("D0-CS-CV", "D1-CS-CV", "D2-CS-CV", "D3-CS-CV"),
                                    "CSCT" = c("D0-CS-CT", "D1-CS-CT", "D2-CS-CT", "D3-CS-CT"),
                                    "CSNT" = c("D0-CS-NT", "D1-CS-NT", "D2-CS-NT", "D3-CS-NT"),
                                  "RCV" = c("D0-R-CV", "D1-R-CV", "D2-R-CV", "D3-R-CV"),
                                  "RCT" = c("D0-R-CT", "D1-R-CT", "D2-R-CT", "D3-R-CT"),
                                  "RNT" = c("D0-R-NT", "D1-R-NT", "D2-R-NT", "D3-R-NT"),
                                  "RR" = c("D0-R-R", "D1-R-R", "D2-R-R", "D3-R-R"))
simulation_shortIDs3 <- data.frame("PDCC" = c("DX-X-CV","PD-2E", "PD-2L", "PD-4L", "PD-6L", "DX-CC-NT"))

simulation_names <- data.frame("Drain0" = c("Drain0", "Drain0CT", "Drain0NT", "RetiredPothole", "ConTillRetired", "NoTillRetired","RetiredAll"),
                               "Drain1" = c("Drain01", "ConTillDrain01", "Drain1NT", "Drain1Retired", "Drain1RetiredCT", "Drain1RetiredNT","Drain1RetiredAll"),
                               "Drain2" = c("Drain02", "ConTillDrain02", "Drain2NT", "Drain2Retired", "Drain2RetiredCT", "Drain2RetiredNT","Drain2RetiredAll"),
                               "Drain3" = c("Drain03", "ConTillDrain03", "Drain3NT", "Drain3Retired", "Drain3RetiredCT", "Drain3RetiredNT","Drain3RetiredAll"))

simulation_names2 <- data.frame("CSCV" = c("Drain0", "Drain01", "Drain02", "Drain03"),
                                "CSCT" = c("Drain0CT", "ConTillDrain01", "ConTillDrain02", "ConTillDrain03"),
                                "CSNT" = c("Drain0NT", "Drain1NT", "Drain2NT", "Drain3NT"),
                                "RCV" = c("RetiredPothole", "Drain1Retired", "Drain2Retired", "Drain3Retired"),
                                "RCT" = c("ConTillRetired", "Drain1RetiredCT", "Drain2RetiredCT", "Drain3RetiredCT"),
                                "RNT" = c("NoTillRetired", "Drain1RetiredNT", "Drain2RetiredNT", "Drain3RetiredNT"),
                                "RR" = c("RetiredAll", "Drain1RetiredAll", "Drain2RetiredAll", "Drain3RetiredAll"))
simulation_names3 <- data.frame("PDCC" = c("Base", "PlantDate01", "PlantDate02", "PlantDate03", "PlantDate04", "CoverCrop"))
```

# Data import

```{r}
max.monthly.by.scen <- read.csv("monthly_maximums.csv", header=T, sep=",")
max.monthly.by.scen <- max.monthly.by.scen[2:6]
```

# Functions

## Plotting monthly maximums

```{r}
# AVERAGE OF MONTHLY MAXIMUM INUNDATION
max.function.partitioned2 <- function(data.df, title, id, supp.labs, labnms) {
  temp <- data.df %>%
    group_by(Month, Pothole, variable) %>%
    summarise(avg.max=mean(value), median.max=median(value))
  
  labs <- as.character(supp.labs)
  names(labs) <- as.character(labnms)
  
  png(filename = paste0("avg.", id, ".comparison.png"), width=1600, height=1000)
  # windows()
  plot.max <- ggplot(temp, aes(x=Month, y=avg.max, fill=Pothole))+
    geom_bar(stat="identity")+
    facet_grid(cols = vars(variable), rows = vars(Pothole), labeller = labeller(variable = labs))+
    ylim(0, 7500)+
    theme(axis.text.x = element_text(angle = 90))+
    theme(text = element_text(size=24))+
    theme(legend.position = "none")+
    labs(
      x="Month",
      y="Max. Volume (m^3)")
  # plot.max
  
  
  return(plot.max)
}

```

## Plotting time series

```{r}
# TIME SERIES 2010-2018
time.series.function.partitioned2 <- function(data.df, title, nm, supp.labs, labnms) {
  
  data.df <- data.df %>%
    filter(Year >= 2010)
  
  labs <- as.character(supp.labs)
  names(labs) <- as.character(labnms)
  
  png(paste("time.",nm,".png"), width=700, height=700)
  # windows()
  plot.anim <- ggplot(data.df, aes(x=Gregorian, y=value, color=variable))+ #,
    geom_line(size=0.5, alpha=0.5) +
    facet_grid(cols = vars(variable), rows = vars(Pothole), labeller = labeller(variable = labs))+
    ylim(0, 50000) +
    theme(legend.position = "none") +
    theme(axis.title.x = element_blank(),
          axis.text.x=element_blank(),
          axis.ticks.x=element_blank(), 
          text=element_text(size=12)) +
    labs(
      # x = "2010",
      y = "Simulated Volume (m^3)",
      title = paste0(title, " , 2010-2018"),
      color="Scenario"
    )
  
  return(plot.anim)
}
```

## Boxplots for average days of inundation

```{r}
# BOX PLOTS OF AVERAGE ANNUAL DAYS OF INUNDATION
boxplot.function.partitioned <- function(data.df, title, nm, supp.labs, labnms) {

  png(paste("days.",nm,".png"), width=700, height=700)
  
plot.avg.annual.by.wetness <- ggplot(data=data.df, aes(x=code, y=value))+
  geom_boxplot(alpha=0.5) +
  facet_grid(rows=vars(wetness),
             labeller = labeller(wetness = c(high = "High Precip", low="Low Precip",
                                             med="Medium Precip"))) +
  theme(axis.text.x =element_text(angle = 90)) +
  theme(text = element_text(size = 12)) +
  ylim(0, 150)+
  labs(
    y = "Average Annual Days of Flooding",
    x = "Simulated Scenario",
    title = title
  )

return(plot.avg.annual.by.wetness)

}

boxplot.function.partitioned.no.wetness <- function(data.df, title, nm, supp.labs, labnms) {
  
  png(paste("days_all_",nm,".png"), width=700, height=700)
  
  plot.avg.annual.by.wetness <- ggplot(data=data.df, aes(x=code, y=value))+
    geom_boxplot(alpha=0.5) +
    # facet_grid(rows=vars(wetness),
    #            labeller = labeller(wetness = c(high = "High Precip", low="Low Precip",
    #                                            med="Medium Precip"))) +
    theme(axis.text.x =element_text(angle = 90)) +
    theme(text = element_text(size = 16)) +
    ylim(0, 150)+
    labs(
      y = "Days Flooded",
      x = "Scenario"#,
      # title = title
    )
  
  return(plot.avg.annual.by.wetness)
  
}
```

# Data

## Maximum monthly

```{r}
max.monthly.by.scen <- max.monthly.by.scen %>%
  spread(variable, value)
```

```{r}
# drainage matrices ====

max.data.drain0 <- max.monthly.by.scen[ , c(1:3, (simulation_numbers$Drain0+4))]
max.data.drain0 <- max.data.drain0 %>%
  gather("variable", "value", 4:10)

max.data.drain0.filter <- max.data.drain0 %>%
  filter(Pothole != "Mouth" & Pothole != "Hen") %>%
  filter(variable == "Drain0" | variable == "RetiredPothole" | variable == "RetiredAll")

max.data.drain1 <- max.monthly.by.scen[ , c(1:3, (simulation_numbers$Drain1+4))]
max.data.drain1 <- max.data.drain1 %>%
  gather("variable", "value", 4:10)

max.data.drain2 <- max.monthly.by.scen[ , c(1:3, (simulation_numbers$Drain2+4))]
max.data.drain2 <- max.data.drain2 %>%
  gather("variable", "value", 4:10)

max.data.drain3 <- max.monthly.by.scen[ , c(1:3, (simulation_numbers$Drain3+4))]
max.data.drain3 <- max.data.drain3 %>%
  gather("variable", "value", 4:10)

# max.data.drain0$facet <- factor(max.data.drain0$variable, levels = c("Drain0", "Drain0CT", "Drain0NT", 
                                                                     # "RetiredPothole", "ConTillRetired", "NoTillRetired",
                                                                     # "RetiredAll"))
```

```{r}
# land use matrices ====
max.data.CSCV <- max.monthly.by.scen[ , c(1:3, (simulation_numbers2$CSCV)+4)]
max.data.CSCV <- max.data.CSCV %>%
  gather("variable", "value", 4:7)

max.data.CSCT <- max.monthly.by.scen[ , c(1:3, (simulation_numbers2$CSCT)+4)]
max.data.CSCT <- max.data.CSCT %>%
  gather("variable", "value", 4:7)

max.data.CSNT <- max.monthly.by.scen[ , c(1:3, (simulation_numbers2$CSNT)+4)]
max.data.CSNT <- max.data.CSNT %>%
  gather("variable", "value", 4:7)

max.data.RCV <- max.monthly.by.scen[ , c(1:3, (simulation_numbers2$RCV)+4)]
max.data.RCV <- max.data.RCV %>%
  gather("variable", "value", 4:7)

max.data.RCT <- max.monthly.by.scen[ , c(1:3, (simulation_numbers2$RCT)+4)]
max.data.RCT <- max.data.RCT %>%
  gather("variable", "value", 4:7)

max.data.RNT <- max.monthly.by.scen[ , c(1:3, (simulation_numbers2$RNT)+4)]
max.data.RNT <- max.data.RNT %>%
  gather("variable", "value", 4:7)

max.data.RR <- max.monthly.by.scen[ , c(1:3, (simulation_numbers2$RR)+4)]
max.data.RR <- max.data.RR %>%
  gather("variable", "value", 4:7)
```

```{r}
# APPLY PLOT FUNCTIONS ======

    # drainage matrices ====

m1 <- max.function.partitioned2(max.data.drain0, "Simulation Comparison, No Drainage Present", "drain0", simulation_shortIDs$Drain0, simulation_names$Drain0)
dev.off()

# max.function.partitioned2(max.data.drain0.filter, "Simulation Comparison, No Drainage Present", "drain0_abbr", simulation_shortIDs$Drain0, simulation_names$Drain0)
# dev.off()

m2 <- max.function.partitioned2(max.data.drain1, "Simulation Comparison, Subsurface Drainage", "drain1", simulation_shortIDs$Drain1, simulation_names$Drain1)
dev.off()

m3 <- max.function.partitioned2(max.data.drain2, "Simulation Comparison, One Inlet", "drain2", simulation_shortIDs$Drain2, simulation_names$Drain2)
dev.off()

m4 <- max.function.partitioned2(max.data.drain3, "Simulation Comparison, Multiple Inlets", "drain3", simulation_shortIDs$Drain3, simulation_names$Drain3)
dev.off()

windows()
# png(filename = "pub_maxplot.png", width=1000, height=750)
ggarrange(m1, m2,m3, m4, 
          labels=c("A", "B", "C", "D"),
          ncol=2, nrow=2)
dev.off()
```


```{r}
max.function.partitioned2(max.data.CSCV, "Simulation Comparison, Conventionally Farmed", "CSCV", simulation_shortIDs2$CSCV, simulation_names2$CSCV)
dev.off()

max.function.partitioned2(max.data.CSCT, "Simulation Comparison, Conservation Tillage", "CSCT", simulation_shortIDs2$CSCT, simulation_names2$CSCT)
dev.off()

max.function.partitioned2(max.data.CSNT, "Simulation Comparison, No Till", "CSNT", simulation_shortIDs2$CSNT, simulation_names2$CSNT)
dev.off()

max.function.partitioned2(max.data.RCV, "Simulation Comparison, Retired Conventional Management", "RCV", simulation_shortIDs2$RCV, simulation_names2$RCV)
dev.off()

max.function.partitioned2(max.data.RCT, "Simulation Comparison, Retired, Conservation Till", "RCT", simulation_shortIDs2$RCT, simulation_names2$RCT)
dev.off()

max.function.partitioned2(max.data.RNT, "Simulation Comparison, Retired, No Till", "RNT", simulation_shortIDs2$RNT, simulation_names2$RNT)
dev.off()

max.function.partitioned2(max.data.RR, "Simulation Comparison, Whole Watershed Retired", "RR", simulation_shortIDs2$RR, simulation_names2$RR)
dev.off()
```

## Annual Days of Inundation

By drainage. 

```{r}
days.data.drain0 <- avg.annual.by.wetness.all.melt %>%
  filter(variable %in% simulation_names$Drain0)

days.data.drain1 <- avg.annual.by.wetness.all.melt %>%
  filter(variable %in% simulation_names$Drain1)

days.data.drain2 <- avg.annual.by.wetness.all.melt %>%
  filter(variable %in% simulation_names$Drain2)

days.data.drain3 <- avg.annual.by.wetness.all.melt %>%
  filter(variable %in% simulation_names$Drain3)
```

By land use practices. 

```{r}
days.data.CSCV <- avg.annual.by.wetness.all.melt %>%
  filter(variable %in% simulation_names2$CSCV)

days.data.CSCT <- avg.annual.by.wetness.all.melt %>%
  filter(variable %in% simulation_names2$CSCT)

days.data.CSNT <- avg.annual.by.wetness.all.melt %>%
  filter(variable %in% simulation_names2$CSNT)

days.data.RCV <- avg.annual.by.wetness.all.melt %>%
  filter(variable %in% simulation_names2$RCV)

days.data.RCT <- avg.annual.by.wetness.all.melt %>%
  filter(variable %in% simulation_names2$RCT)

days.data.RNT <- avg.annual.by.wetness.all.melt %>%
  filter(variable %in% simulation_names2$RNT)

days.data.RR <- avg.annual.by.wetness.all.melt %>%
  filter(variable %in% simulation_names2$RR)
```

## Applying the functions

```{r}
boxplot.function.partitioned(days.data.drain0, "Simulation Comparison, No Drainage Present", "drain0", simulation_shortIDs$Drain0, simulation_names$Drain0)
dev.off()

boxplot.function.partitioned(days.data.drain1, "Simulation Comparison, Subsurface Drainage", "drain1", simulation_shortIDs$Drain1, simulation_names$Drain1)
dev.off()

boxplot.function.partitioned(days.data.drain2, "Simulation Comparison, One Inlet", "drain2", simulation_shortIDs$Drain2, simulation_names$Drain2)
dev.off()

boxplot.function.partitioned(days.data.drain3, "Simulation Comparison, Multiple Inlets", "drain3", simulation_shortIDs$Drain3, simulation_names$Drain3)
dev.off()
```


```{r}
boxplot.function.partitioned(days.data.CSCV, "Simulation Comparison, Conventionally Farmed", "CSCV", simulation_shortIDs2$CSCV, simulation_names2$CSCV)
dev.off()

boxplot.function.partitioned(days.data.CSCT, "Simulation Comparison, Conservation Tillage", "CSCT", simulation_shortIDs2$CSCT, simulation_names2$CSCT)
dev.off()

boxplot.function.partitioned(days.data.CSNT, "Simulation Comparison, No Till", "CSNT", simulation_shortIDs2$CSNT, simulation_names2$CSNT)
dev.off()

boxplot.function.partitioned(days.data.RCV, "Simulation Comparison, Retired Conventional Management", "RCV", simulation_shortIDs2$RCV, simulation_names2$RCV)
dev.off()

boxplot.function.partitioned(days.data.RCT, "Simulation Comparison, Retired, Conservation Till", "RCT", simulation_shortIDs2$RCT, simulation_names2$RCT)
dev.off()

boxplot.function.partitioned(days.data.RNT, "Simulation Comparison, Retired, No Till", "RNT", simulation_shortIDs2$RNT, simulation_names2$RNT)
dev.off()

boxplot.function.partitioned(days.data.RR, "Simulation Comparison, Whole Watershed Retired", "RR", simulation_shortIDs2$RR, simulation_names2$RR)
dev.off()
```

### Other

```{r}
# SECONDARY BOXPLOTS (NOT BY WETNESS) =====


p1 <- boxplot.function.partitioned.no.wetness(days.data.drain0, "Simulation Comparison, No Drainage Present", "drain0", simulation_shortIDs$Drain0, simulation_names$Drain0)
dev.off()

p2 <- boxplot.function.partitioned.no.wetness(days.data.drain1, "Simulation Comparison, Subsurface Drainage", "drain1", simulation_shortIDs$Drain1, simulation_names$Drain1)
dev.off()

p3 <- boxplot.function.partitioned.no.wetness(days.data.drain2, "Simulation Comparison, One Inlet", "drain2", simulation_shortIDs$Drain2, simulation_names$Drain2)
dev.off()

p4 <- boxplot.function.partitioned.no.wetness(days.data.drain3, "Simulation Comparison, Multiple Inlets", "drain3", simulation_shortIDs$Drain3, simulation_names$Drain3)
dev.off()

png(filename = "pub_bxplot.png", width=1000, height=750)
ggarrange(p1, p2,p3, p4, 
          labels=c("A", "B", "C", "D"),
          ncol=2, nrow=2)
dev.off()

boxplot.function.partitioned.no.wetness(days.data.CSCV, "Simulation Comparison, Conventionally Farmed", "CSCV", simulation_shortIDs2$CSCV, simulation_names2$CSCV)
dev.off()

boxplot.function.partitioned.no.wetness(days.data.CSCT, "Simulation Comparison, Conservation Tillage", "CSCT", simulation_shortIDs2$CSCT, simulation_names2$CSCT)
dev.off()

boxplot.function.partitioned.no.wetness(days.data.CSNT, "Simulation Comparison, No Till", "CSNT", simulation_shortIDs2$CSNT, simulation_names2$CSNT)
dev.off()

boxplot.function.partitioned.no.wetness(days.data.RCV, "Simulation Comparison, Retired Conventional Management", "RCV", simulation_shortIDs2$RCV, simulation_names2$RCV)
dev.off()

boxplot.function.partitioned.no.wetness(days.data.RCT, "Simulation Comparison, Retired, Conservation Till", "RCT", simulation_shortIDs2$RCT, simulation_names2$RCT)
dev.off()

boxplot.function.partitioned.no.wetness(days.data.RNT, "Simulation Comparison, Retired, No Till", "RNT", simulation_shortIDs2$RNT, simulation_names2$RNT)
dev.off()

boxplot.function.partitioned.no.wetness(days.data.RR, "Simulation Comparison, Whole Watershed Retired", "RR", simulation_shortIDs2$RR, simulation_names2$RR)
dev.off()

boxplot.function.partitioned.no.wetness(days.data.PDCC, "Simulation Comparison, Planting Date and cover Crop", "PDCC", simulation_shortIDs3$PDCC, simulation_names3$PDCC)
dev.off()
```
