==> 35_36_Ch3_FigAreaFreq.Rmd <==
---
title: "Publication Figure Script"
author: "Brady Nahkala"
date: "9/1/2020"
output: html_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
```

# Figure 

# Upload data

```{r}
area.bunny.df <- read.csv("area_time_series_bunny.csv", header=T, sep=",")
area.walnut.df <- read.csv("area_time_series_walnut.csv", header=T, sep=",")
area.lettuce.df <- read.csv("area_time_series_lettuce.csv", header=T, sep=",")
area.cardinal.df <- read.csv("area_time_series_cardinal.csv", header=T, sep=",")
area.hen.df <- read.csv("area_time_series_hen.csv", header=T, sep=",")
area.mouth.df <- read.csv("area_time_series_mouth.csv", header=T, sep=",")
```

# Script

```{r}
# LIBRARY =====
library(dplyr)
library(tidyr)
library(ggplot2)
library(gridExtra)
library(ggpubr)
library(purrr)
library(stringr)
```


```{r, fig.width=7.09, fig.height=7.09}



# NAME METADATA ====
simulation_shortIDs <- data.frame("Drain0" = c("D0-CS-CV", "D0-CS-CT", "D0-CS-NT", "D0-R-CV", "D0-R-CT", "D0-R-NT","D0-R-R"),
                                  "Drain1" = c("D1-CS-CV", "D1-CS-CT", "D1-CS-NT", "D1-R-CV", "D1-R-CT", "D1-R-NT","D1-R-R"),
                                  "Drain2" = c("D2-CS-CV", "D2-CS-CT", "D2-CS-NT", "D2-R-CV", "D2-R-CT", "D2-R-NT","D2-R-R"),
                                  "Drain3" = c("D3-CS-CV", "D3-CS-CT", "D3-CS-NT", "D3-R-CV", "D3-R-CT", "D3-R-NT","D3-R-R"))
simulation_shortIDs2 <- data.frame("CSCV" = c("D0-CS-CV", "D1-CS-CV", "D2-CS-CV", "D3-CS-CV"),
                                   "CSCT" = c("D0-CS-CT", "D1-CS-CT", "D2-CS-CT", "D3-CS-CT"),
                                   "CSNT" = c("D0-CS-NT", "D1-CS-NT", "D2-CS-NT", "D3-CS-NT"),
                                   "RCV" = c("D0-R-CV", "D1-R-CV", "D2-R-CV", "D3-R-CV"),
                                   "RCT" = c("D0-R-CT", "D1-R-CT", "D2-R-CT", "D3-R-CT"),
                                   "RNT" = c("D0-R-NT", "D1-R-NT", "D2-R-NT", "D3-R-NT"),
                                   "RR" = c("D0-R-R", "D1-R-R", "D2-R-R", "D3-R-R"))
simulation_shortIDs3 <- data.frame("PDCC" = c("DX-X-CV","PD-2E", "PD-2L", "PD-4L", "PD-6L", "DX-CC-NT"))

ids <- simulation_shortIDs %>%
  gather(key="type", "id", 1:4)
ids2 <- simulation_shortIDs2 %>%
  gather(key="type", "id", 1:7)
ids3 <- simulation_shortIDs3 %>%
  gather(key="type", "id", 1)

ids <- ids %>%
  bind_rows(ids3) 

simulation_names <- data.frame("Drain0" = c("Drain0", "Drain0CT", "Drain0NT", "RetiredPothole", "ConTillRetired", "NoTillRetired","RetiredAll"),
                               "Drain1" = c("Drain01", "ConTillDrain01", "Drain1NT", "Drain1Retired", "Drain1RetiredCT", "Drain1RetiredNT","Drain1RetiredAll"),
                               "Drain2" = c("Drain02", "ConTillDrain02", "Drain2NT", "Drain2Retired", "Drain2RetiredCT", "Drain2RetiredNT","Drain2RetiredAll"),
                               "Drain3" = c("Drain03", "ConTillDrain03", "Drain3NT", "Drain3Retired", "Drain3RetiredCT", "Drain3RetiredNT","Drain3RetiredAll"))
simulation_names2 <- data.frame("CSCV" = c("Drain0", "Drain01", "Drain02", "Drain03"),
                                "CSCT" = c("Drain0CT", "ConTillDrain01", "ConTillDrain02", "ConTillDrain03"),
                                "CSNT" = c("Drain0NT", "Drain1NT", "Drain2NT", "Drain3NT"),
                                "RCV" = c("RetiredPothole", "Drain1Retired", "Drain2Retired", "Drain3Retired"),
                                "RCT" = c("ConTillRetired", "Drain1RetiredCT", "Drain2RetiredCT", "Drain3RetiredCT"),
                                "RNT" = c("NoTillRetired", "Drain1RetiredNT", "Drain2RetiredNT", "Drain3RetiredNT"),
                                "RR" = c("RetiredAll", "Drain1RetiredAll", "Drain2RetiredAll", "Drain3RetiredAll"))
simulation_names3 <- data.frame("PDCC" = c("Base", "PlantDate01", "PlantDate02", "PlantDate03", "PlantDate04", "CoverCrop"))

nms <- simulation_names %>%
  gather(key="type", "name", 1:4)
nms2 <- simulation_names2 %>%
  gather(key="type", "name", 1:7)
nms3 <- simulation_names3 %>%
  gather(key="type", "name", 1)

nms <- nms %>%
  bind_rows(nms3) 




# VIZ ====

# area.bunny.df.filter <- area.walnut.df %>%
#   # filter(Base != 0) %>%
#   gather(key="scen", value="area", 5:43) %>%
#   group_by(Year, scen) %>%
#   summarize(pct.max.year = (max(area)/(chars.potholes$area.max[1]*10000))) #%>%
#   # filter(Month >= 4 & Month <= 9) #%>%
#   # summarize(distr=n())

# GENERATE DATA ====
pct.area.data.function <- function(data.df, id) {
  
  temp <- data.df %>%
    # filter(Base != 0) %>%
    gather(key="scen", value="area", 5:43) %>%
    group_by(Year, scen) %>%
    summarize(pct.max.year = (max(area)/(chars.potholes$area.max[id]*10000))) #%>%
  # filter(Month >= 4 & Month <= 9) #%>%
  # summarize(distr=n())
  
  
  zero <- temp %>%
    group_by(scen) %>%
    filter(pct.max.year >= 0) %>%
    summarize(zero = n()/25)
  
  one <- temp %>%
    group_by(scen) %>%
    filter(pct.max.year > 0.1) %>%
    summarize(one = n()/25)
  
  two <- temp %>%
    group_by(scen) %>%
    filter(pct.max.year > 0.2) %>%
    summarize(two = n()/25)
  
  three <- temp %>%
    group_by(scen) %>%
    filter(pct.max.year > 0.3) %>%
    summarize(three = n()/25)
  
  four <- temp %>%
    group_by(scen) %>%
    filter(pct.max.year > 0.4) %>%
    summarize(four = n()/25)
  
  five <- temp %>%
    group_by(scen) %>%
    filter(pct.max.year > 0.5) %>%
    summarize(five = n()/25)
  
  six <- temp %>%
    group_by(scen) %>%
    filter(pct.max.year > 0.6) %>%
    summarize(six = n()/25)
  
  seven <- temp %>%
    group_by(scen) %>%
    filter(pct.max.year > 0.7) %>%
    summarize(seven = n()/25)
  
  eight <- temp %>%
    group_by(scen) %>%
    filter(pct.max.year > 0.8) %>%
    summarize(eight = n()/25)
  
  nine <- temp %>%
    group_by(scen) %>%
    filter(pct.max.year > 0.9) %>%
    summarize(nine = n()/25)
  
  ten <- temp %>%
    group_by(scen) %>%
    filter(pct.max.year >= 1.0) %>%
    summarize(ten = n()/25)
  
  all <- merge(zero, one, by="scen", all=TRUE)
  all <- merge(all, two, by="scen", all=TRUE)
  all <- merge(all, three, by="scen", all=TRUE)
  all <- merge(all, four, by="scen", all=TRUE)
  all <- merge(all, five, by="scen", all=TRUE)
  all <- merge(all, six, by="scen", all=TRUE)
  all <- merge(all, seven, by="scen", all=TRUE)
  all <- merge(all, eight, by="scen", all=TRUE)
  all <- merge(all, nine, by="scen", all=TRUE)
  all <- merge(all, ten, by="scen", all=TRUE)
  
  colnames(all) <- c("scen", "zero", "one", "two", "three", "four", "five", "six", "seven", "eight", "nine", "ten")
  
  all.melt <- all %>%
    gather(key = "range", value="pct", 2:12)
  
  all.melt$area.pct <- c(rep(0, 39), rep(0.1, 39), rep(0.2, 39),rep(0.3, 39),rep(0.4, 39),rep(0.5, 39),rep(0.6, 39),
                         rep(0.7, 39),rep(0.8, 39),rep(0.9, 39),rep(1, 39))
  
  all.melt$pct[is.na(all.melt$pct)] <- 0
  
  return(all.melt)
  
}

area.curve.bunny <- pct.area.data.function(area.bunny.df, 1)
area.curve.walnut <- pct.area.data.function(area.walnut.df, 2)
area.curve.lettuce <- pct.area.data.function(area.lettuce.df, 3)
area.curve.cardinal <- pct.area.data.function(area.cardinal.df, 4)
area.curve.hen <- pct.area.data.function(area.hen.df, 5)
area.curve.mouth <- pct.area.data.function(area.mouth.df, 6)

area.curve.all <- area.curve.bunny %>%
  bind_rows(area.curve.walnut) %>%
  bind_rows(area.curve.lettuce) %>%
  bind_rows(area.curve.cardinal) %>%
  bind_rows(area.curve.hen) %>%
  bind_rows(area.curve.mouth) 

area.curve.all$Pothole <- c(rep("Bunny", 429),rep("Walnut", 429),rep("Lettuce", 429),
                            rep("Cardinal", 429),rep("Hen", 429),rep("Mouth", 429))

area.curve.all <- area.curve.all %>%
  filter(scen %in% nms$name)

# SOME DATA SPLITTING ====

# drainage =====
area.curve.drain0 <- area.curve.all %>%
  filter(scen %in% nms$name[nms$type == "Drain0"])

area.curve.drain0$scen2 <- factor(area.curve.drain0$scen, levels = c("Drain0", "Drain0CT", "Drain0NT",
                                                                     "RetiredPothole", "ConTillRetired",
                                                                     "NoTillRetired", "RetiredAll"))

area.curve.drain1 <- area.curve.all %>%
  filter(scen %in% nms$name[nms$type == "Drain1"])

area.curve.drain1$scen2 <- factor(area.curve.drain1$scen, levels=c("Drain01", "ConTillDrain01", "Drain1NT",
                                                                   "Drain1Retired", "Drain1RetiredCT", 
                                                                   "Drain1RetiredNT", "Drain1RetiredAll"))

area.curve.drain2 <- area.curve.all %>%
  filter(scen %in% nms$name[nms$type == "Drain2"])

area.curve.drain2$scen2 <- factor(area.curve.drain2$scen, levels=c("Drain02", "ConTillDrain02", "Drain2NT",
                                                                   "Drain2Retired", "Drain2RetiredCT", 
                                                                   "Drain2RetiredNT", "Drain2RetiredAll"))


area.curve.drain3 <- area.curve.all %>%
  filter(scen %in% nms$name[nms$type == "Drain3"])

area.curve.drain3$scen2 <- factor(area.curve.drain3$scen, levels=c("Drain03", "ConTillDrain03", "Drain3NT",
                                                                   "Drain3Retired", "Drain3RetiredCT", 
                                                                   "Drain3RetiredNT", "Drain3RetiredAll"))

area.curves.by.drain <- area.curve.drain0 %>%
  bind_rows(area.curve.drain1) %>%
  bind_rows(area.curve.drain2) %>%
  bind_rows(area.curve.drain3) 
area.curves.by.drain$group <- c(rep("drain0", 462),rep("drain1", 462),rep("drain2", 462),rep("drain3", 462))  

write.csv(area.curve.all, file = "area_curves.csv")


# land use  =====
area.curve.CSCV <- area.curve.all %>%
  filter(scen %in% nms2$name[nms2$type == "CSCV"])

area.curve.CSCT <- area.curve.all %>%
  filter(scen %in% nms2$name[nms2$type == "CSCT"])

area.curve.CSNT <- area.curve.all %>%
  filter(scen %in% nms2$name[nms2$type == "CSNT"])

area.curve.RCV <- area.curve.all %>%
  filter(scen %in% nms2$name[nms2$type == "RCV"])

area.curve.RCT <- area.curve.all %>%
  filter(scen %in% nms2$name[nms2$type == "RCT"])

area.curve.RNT <- area.curve.all %>%
  filter(scen %in% nms2$name[nms2$type == "RNT"])

area.curve.RR <- area.curve.all %>%
  filter(scen %in% nms2$name[nms2$type == "RR"])

area.curves.by.LU <- area.curve.CSCV %>%
  bind_rows(area.curve.CSCT) %>%
  bind_rows(area.curve.CSNT) %>%
  bind_rows(area.curve.RCV) %>%
  bind_rows(area.curve.RCT) %>%
  bind_rows(area.curve.RNT) %>%
  bind_rows(area.curve.RR) 
area.curves.by.LU$group <-
  c(
    rep("CSCV", 264),
    rep("CSCT", 264),
    rep("CSNT", 264),
    rep("RCV", 264),
    rep("RCT", 264),
    rep("RNT", 264),
    rep("RR", 264)
  )


# PLOT DATA =====
plot.area.curves <- function(data.df, nm, supp.labs, labnms) {
  
  labs <- as.character(supp.labs)
  names(labs) <- as.character(labnms)
  
  plt <- ggplot(data.df, aes(x=pct, y=area.pct, color=scen2))+
    geom_path(linetype="solid", size=0.6)+
    facet_grid(rows = vars(Pothole))+
    labs(
      x="",
      y="",
      color=""
    )+
    theme_bw(base_size=9)+
    theme(legend.position = "none")+
    scale_color_discrete(name="Simulation Scenario", labels=labs)+
    # scale_y_continuous(breaks = c("0", "0.5",  "1"))+
    scale_color_brewer(palette = "Dark2", labels=c("DX-CS-CV","DX-CS-CT","DX-CS-NT","DX-R-CV","DX-R-CT","DX-R-NT", "DX-R-R"))+
    # ylim(0, 1.1)+
    xlim(0,1)
  
  
  png(filename=paste0("area_curves_", nm,".png"), width=600, height=600)
  # windows()
  plt
}

# PLOT PARTITIONED SCENARIOS =====
g1 <- plot.area.curves(area.curve.drain0, "drain0", simulation_shortIDs$Drain0, simulation_names$Drain0)
g1 <- g1 + theme(axis.text.x = element_blank())
dev.off()
g2 <- plot.area.curves(area.curve.drain1, "drain1", simulation_shortIDs$Drain1, simulation_names$Drain1)
g2 <- g2 + theme(axis.text.x = element_blank()) + theme(axis.text.y = element_blank())
dev.off() 
g3 <- plot.area.curves(area.curve.drain2, "drain2", simulation_shortIDs$Drain2, simulation_names$Drain2)
dev.off() 
g4 <- plot.area.curves(area.curve.drain3, "drain3", simulation_shortIDs$Drain3, simulation_names$Drain3)
g4 <- g4  + theme(axis.text.y = element_blank())
dev.off()


# using ggplots generated in data_viz.R


png(filename = "Fig5_AreaFrequency.png", height = 190, width = 190, units = "mm", res = 300)

arr <- ggarrange(
  g1,
  g2,
  g3,
  g4,
  labels = c("A", "B", "C", "D"),
  ncol = 2,
  nrow = 2, 
  common.legend = T, 
  legend = "top"
)
annotate_figure(arr, 
                left =text_grob("Fraction of Maximum Pothole Volume", rot = 90),
                bottom = text_grob("Fraction of Years Simulated"))

dev.off()
```


# Attempt at Figure 5 Revision, based on Response to Review, Reviewer 1

```{r, fig.width=7.09, fig.height=7.09}
# area.curve.drain0$drain <- "D0"
# area.curve.drain1$drain <- "D1"
# area.curve.drain2$drain <- "D2"
# area.curve.drain3$drain <- "D3"
# 
# area.curve.drain <- rbind(area.curve.drain0, area.curve.drain1, area.curve.drain2, area.curve.drain3)
# 
# area.curve.drain$drain <- factor(area.curve.drain$drain, levels = c("D0", "D1", "D2", "D3"))

# area.curve.drain$scen2 <- factor(area.curve.drain$scen, levels = c("Drain0", "Drain0CT", "Drain0NT",
#                                                                      "RetiredPothole", "ConTillRetired",
#                                                                      "NoTillRetired", "RetiredAll", 
#                                                                     "Drain01", "ConTillDrain01", "Drain1NT",
#                                                                    "Drain1Retired", "Drain1RetiredCT", 
#                                                                    "Drain1RetiredNT", "Drain1RetiredAll", 
#                                                                    "Drain02", "ConTillDrain02", "Drain2NT",
#                                                                    "Drain2Retired", "Drain2RetiredCT", 
#                                                                    "Drain2RetiredNT", "Drain2RetiredAll", 
#                                                                    "Drain03", "ConTillDrain03", "Drain3NT",
#                                                                    "Drain3Retired", "Drain3RetiredCT", 
#                                                                    "Drain3RetiredNT", "Drain3RetiredAll"))


plot.area.curves_Revised <- function(data.df, nm, supp.labs, labnms) {
  
  labs <- as.character(supp.labs)
  names(labs) <- as.character(labnms)
  
  plt <- ggplot(data.df, aes(x=pct, y=area.pct, color=scen2, linetype=drain))+
    geom_path(size=0.6)+
    facet_grid(rows = vars(Pothole))+
    labs(
      x="Fraction of Years Simulation",
      y="Fraction of Maximum Pothole Volume",
      color="",
      linetype=""
    )+
    theme_bw(base_size=9)+
    theme(legend.position = "top")+
    scale_color_discrete(name="Simulation Scenario", labels=labs)+
    # scale_y_continuous(breaks = c("0", "0.5",  "1"))+
    scale_color_brewer(palette = "Dark2", labels=c("DX-CS-CV","DX-CS-CT","DX-CS-NT","DX-R-CV","DX-R-CT","DX-R-NT", "DX-R-R"))+
    # ylim(0, 1.1)+
    xlim(0,1)
  
  
  png(filename=paste0("area_curves_", nm,".png"), width=600, height=600)
  # windows()
  plt
}

g_Rev <- plot.area.curves_Revised(area.curve.drain, "drain", simulation_shortIDs$Drain1, simulation_names$Drain1)

png(filename = "Fig5_AreaFrequency_Revised.png", height = 190, width = 190, units = "mm", res = 300)
g_Rev
dev.off()

g_Rev

dev.off()
```




# Loess smoothing

```{r}
area.curves.by.drain.b<- area.curves.by.drain %>%
  filter(Pothole == chars.potholes$pothole[1])
area.curves.by.drain.w <- area.curves.by.drain %>%
  filter(Pothole == chars.potholes$pothole[2])
area.curves.by.drain.l <- area.curves.by.drain %>%
  filter(Pothole == chars.potholes$pothole[3])
area.curves.by.drain.c <- area.curves.by.drain %>%
  filter(Pothole == chars.potholes$pothole[4])
area.curves.by.drain.h <- area.curves.by.drain %>%
  filter(Pothole == chars.potholes$pothole[5])
area.curves.by.drain.m <- area.curves.by.drain %>%
  filter(Pothole == chars.potholes$pothole[6])

area.curves.by.drain.tst <- area.curves.by.drain %>%
  filter(Pothole != chars.potholes$pothole[6])
```

```{r, fig.width=3.54, fig.height=4}
# FIG TESTING S4.4 ====
Fig7a <- ggplot(area.curves.by.drain.tst, aes(x=pct, y=area.pct, color=group))+
  geom_smooth(method = "loess", se=TRUE, alpha=0.35)+
  theme_bw(base_size = 9)+
  scale_color_discrete(name=NULL,
                       breaks=c("drain0", "drain1", "drain2", "drain3"),
                       labels=c("D0-X-X", "D1-X-X",
                                "D2-X-X", "D3-X-X"))+
  labs(
    x=NULL,
    y="Percent of Total Pothole Area"#,
    # color="Drainage Status"#,
    # title="All potholes but Mouth"
  )+
  xlim(0, 1)+
  ylim(0,1.1)+
  theme(legend.position = "top")+
  guides(color=guide_legend(nrow=2))
  

# png(filename = "loess_tst_fit_to_all_scen_by_pothole.png", width=800, height=800)
# windows()
Fig7a
# dev.off()
```

```{r, fig.width=3.54, fig.height=4}
area.curves.by.LU.tst <- area.curves.by.LU %>%
  filter(Pothole != chars.potholes$pothole[6])

Fig7b <- ggplot(area.curves.by.LU.tst, aes(x=pct, y=area.pct, color=group))+
  geom_smooth(method = "loess", se=T, alpha=0.35)+
  theme_bw(base_size = 9)+
  scale_color_discrete(name=NULL,
                       breaks=c("CSCV", "CSCT", "CSNT", "RCV", "RCT", "RNT", "RR"),
                       labels=c("DX-CS-CV", "DX-CS-CT",
                                "DX-CS-NT", "DX-R-CV",
                                "DX-R-CT", "DX-R-NT",
                                "DX-R-R"))+
  # scale_color_discrete(labels=function(x) str_wrap(x, width = 1))+
  labs(
    x=NULL,
    y=NULL#,
    # color="Drainage Status"#,
    # title="All potholes but Mouth"
  )+
  xlim(0, 1)+
  ylim(0,1.1)+
  theme(legend.position = "top")+
  guides(color=guide_legend(nrow=2))

# png(filename = "loess_tst_fit_to_all_scen_by_LU.png", width=800, height=800)
# windows()
Fig7b
# dev.off()
```

```{r, fig.width=7.1, fig.height=4}
Fig7_Loess <- ggarrange(Fig7a, Fig7b, labels = c("A", "B"))
annotate_figure(Fig7_Loess, bottom = text_grob("Percent of Simulated Years", size=9))

png("Fig7_AFbyFactor.png", width = 190, height=102, units = "mm", res = 300)
Fig7_Loess
annotate_figure(Fig7_Loess, bottom = text_grob("Percent of Simulated Years", size=9))
dev.off()
```
