==> 43_Ch4_FigAreas.Rmd <==
---
title: "ResultsFig ML Pub"
author: "Brady Nahkala"
date: "6/26/2020"
output: html_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
```

```{r}
library(tidyr)
library(dplyr)
library(ggplot2)
library(reshape2)
library(ggpubr)
```

```{r}
maxareaML <- read.csv("CH4_MaxArea.csv", header=T, sep=",")
```


```{r}
# maxareaML <- rbind(max.data.drain0, max.data.drain1, max.data.drain2, max.data.drain3)
maxareaML$scen <- factor(maxareaML$scen)

head(maxareaML)
```

```{r}
maxareaML <- maxareaML %>%
    group_by(Month, Pothole, scen) %>%
    summarise(avg.max=mean(value), median.max=median(value)) #%>%
    # filter(avg.max != 0)

maxareaML$Month <- factor(temp$Month)
head(maxareaML)
# maxareaML$avg.max[maxareaML$Pothole=="Bunny"] <- maxareaML$avg.max / (chars.potholes$area.max[1] * 10000)
# maxareaML$avg.max[maxareaML$Pothole=="Walnut"] <- maxareaML$avg.max / (chars.potholes$area.max[2] * 10000)
# maxareaML$avg.max[maxareaML$Pothole=="Lettuce"] <- maxareaML$avg.max / (chars.potholes$area.max[3] * 10000)
# maxareaML$avg.max[maxareaML$Pothole=="Cardinal"] <- maxareaML$avg.max / (chars.potholes$area.max[4] * 10000)
# maxareaML$avg.max[maxareaML$Pothole=="Hen"] <- maxareaML$avg.max / (chars.potholes$area.max[5] * 10000)
# maxareaML$avg.max[maxareaML$Pothole=="Mouth"] <- maxareaML$avg.max / (chars.potholes$area.max[6] * 10000)
```

```{r}
pub4.1 <- ggplot(maxareaML, aes(x = Month, y = avg.max)) +
  geom_boxplot() +
  facet_grid(cols = vars(Pothole)) +
  theme(text = element_text(size = 12)) +
  theme(legend.position = "none") +
  ylim(0, 25000) +
  # scale_y_discrete(limits = c(0, 10000, 20000)) +
  labs(x = "Month",
       y = "Surface Area (m^2)")

pub4.1
```



```{r}
png("pubML4.1.png", width = 800, height=400)
pub4.1
dev.off()
```

```{r}
pub4.1data <- layer_data(pub4.1, 1)
```
